data binding of components
all components have exactly one field, and can have any services
injected
there is a setter for the field, written in fat arrow notation
the field can be associated with a repository event of the type of the
field. in this case the setter is subscribed to the event, and the
repository is injected

the services can be used to communicate data changes to the
environment

this way the component code can be generated from the model, the state
of the component is just and implementation detail (not master, driven
by services), and no mechanisms wlowing down with unnecessary events.

tldr: there is one field, it gets updated by a repository event, and the
changes are communicated by services.